New members are now sent a DM when joining, also... moved two old listeners into a single class in the listener package to organize them.
