+++ src/main/kotlin/com/learnspigot/bot/LearnSpigotBot.kt
@@ -3,10 +3,9 @@ package com.learnspigot.bot
 import com.learnspigot.bot.command.*
 import com.learnspigot.bot.dbmigrator.migrations.UserProfileMigrator
 import com.learnspigot.bot.entity.DataFile
-import com.learnspigot.bot.entity.SerializedMessage
 import com.learnspigot.bot.entity.UserProfile
 import com.learnspigot.bot.http.UdemyService
-import com.learnspigot.bot.listener.SuggestionChannelListener
+import com.learnspigot.bot.listener.UserListeners
 import com.learnspigot.bot.manager.*
 import com.learnspigot.bot.util.LectureSearcher
 import com.mongodb.client.MongoClients
@@ -26,7 +25,6 @@ import net.dv8tion.jda.api.entities.Guild
 import net.dv8tion.jda.api.entities.Message
 import net.dv8tion.jda.api.entities.User
 import net.dv8tion.jda.api.events.interaction.command.GenericCommandInteractionEvent
-import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import net.dv8tion.jda.api.interactions.callbacks.IReplyCallback
 import net.dv8tion.jda.api.interactions.components.LayoutComponent
 import net.dv8tion.jda.api.requests.GatewayIntent
@@ -90,12 +88,6 @@ class LearnSpigotBot {
         pollManager = PollManager(bot)
         registerCommands()
         registerListeners()
-        bot.listener<MessageReceivedEvent> {
-            if(it.member == null) return@listener
-            val profile: UserProfile = datastore.findUserProfile(it.member!!.id)
-            profile.messageHistory.add(SerializedMessage.fromDiscordMessage(it.message))
-            datastore.save(profile)
-        }
 
         GlobalScope.launch {
             activityJob = async {
@@ -177,7 +169,8 @@ class LearnSpigotBot {
     }
 
     private fun registerListeners() {
-        SuggestionChannelListener(bot)
+        logger.info("Registering listeners...")
+        UserListeners(guild, bot, datastore)
     }
